Implement logic for new difficutly levels.
When mastery is achieved, meaning the user has had 4 out of 5 correct
answeres for a given level, there is now logic to advance the user to
the next level. There are currently three levels of difficulty which
have their own respective attributes. When a user quits the game, their
highest level achieved will now be recorded to the database.